+++ index.ts
@@ -32,10 +32,7 @@ export {
   default as AbstractFeatureProvider,
   AbstractFeatureProviderOptions,
 } from './src/featureProvider/abstractFeatureProvider.js';
-export {
-  isProvidedFeature,
-  showProvidedFeature,
-} from './src/featureProvider/featureProviderSymbols.js';
+export { isProvidedFeature } from './src/featureProvider/featureProviderSymbols.js';
 export {
   default as TileProviderFeatureProvider,
   TileProviderFeatureProviderOptions,