+++ src/featureProvider/featureProviderSymbols.ts
@@ -1,10 +1,5 @@
 /**
  * Added to ol.Feature, if they are not part of a layer, but provided by an {@link AbstractFeatureProvider}.
  */
+// eslint-disable-next-line import/prefer-default-export
 export const isProvidedFeature: unique symbol = Symbol('isProvidedFeature');
-
-/**
- * A boolean value, indicating whether selecting should add the feature to the selected item layer
- * @deprecated
- */
-export const showProvidedFeature: unique symbol = Symbol('showProvidedFeature');