+++ src/layer/layerSymbols.ts
@@ -1,2 +1,5 @@
+/**
+ * Symbol to declare a layers name on its visualizations, e.g. ol.layer.Layer, Cesium.Cesium3DTileset*
+ */
 /* eslint-disable import/prefer-default-export */
 export const vcsLayerName: unique symbol = Symbol('vcsLayerName');